Fix SIGNEXT macro to return correct values for negative coordinates

+++ mcp_gamestate.h
@@ -84,7 +84,7 @@ typedef struct {
 #define CC_2(X,Z)   (uint32_t)(((((uint64_t)(X))>>13)&0x1ff)|(((((uint64_t)(Z))>>13)&0x1ff)<<9))
 
 // offsets in world, super-region, region -> chunk coord
-#define SIGNEXT(X) ((X)|(((X)&0x200000)?0xffc00000:0))
+#define SIGNEXT(X) ((int32_t)((X)|(((X)&0x200000)?0xffc00000:0)))
 #define CC_X(S,R,C) SIGNEXT(  (((S)&0x1ff)<<13)  | (((R)&0xff)<<5)   | ((C)&0x1f)        )
 #define CC_Z(S,R,C) SIGNEXT(  (((S)&0x3fe00)<<4) | (((R)&0xff00)>>3) | (((C)&0x3e0)>>5)  )
